+++ golang/lib/binding_constructors/binding_constructors.go
@@ -64,8 +64,10 @@ func NewRegisterServiceArgs(serviceId string, partitionId string) *kurtosis_core
 	}
 }
 
-func NewRegisterServiceResponse(ipAddr string) *kurtosis_core_rpc_api_bindings.RegisterServiceResponse {
-	return &kurtosis_core_rpc_api_bindings.RegisterServiceResponse{IpAddr: ipAddr}
+func NewRegisterServiceResponse(relativeServiceDirpath string) *kurtosis_core_rpc_api_bindings.RegisterServiceResponse {
+	return &kurtosis_core_rpc_api_bindings.RegisterServiceResponse{
+		RelativeServiceDirpath: relativeServiceDirpath,
+	}
 }
 
 // ==============================================================================================