+++ docs/changelog.md
@@ -1,4 +1,6 @@
 # TBD
+
+# 0.17.2
 ### Features
 * Added `StartExternalContainerRegistration` and `FinishExternalContainerRegistration` endpoints, for registering containers started by external sources into the enclave